Enhance messaging feature with ride context and real-time chat
